% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CI.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculate confidence}
\usage{
calculate_CI(obj, ...)
}
\arguments{
\item{x}{the input data vector with the length of N}

\item{conf}{the \code{1-α} confidence interval}
}
\value{
The lower and upper bound in the confidence interval
}
\description{
Calculate the confidence intervals from simulated data
}
\details{
This function calculates the confidence intervals of a sample mean and returns a named vector of length 2, where the first value is the lower_bound, the second value is the upper_bound.

If the input data contains NaN value, the function will print warning to remind.

If the input is empty, the function will print warning to remind

If the input doesn't belong to numeric or ci_class class, the function will print warning to remind.

If the 1-α lower than 0 or larger than 1, the function will print warning to remind.
}
\examples{
calculate_CI(c(1:10), 0.95)
calculate_CI(c(1:10, 'String'), 0.95)

library(SunanP2)
set.seed(1234)
x <- rnorm(100)
obj <- make_ci_class(x)
output = calculate_CI(obj, 0.95)
}
