% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sin.R
\name{fn_sin}
\alias{fn_sin}
\title{Sine function}
\usage{
fn_sin(x, k = 1)
}
\arguments{
\item{x}{the number to be transformed}

\item{k}{the number of terms to be used in the series expansion beyond the constant 1. The value of k is always over than 1.}
}
\value{
The approximation to the sine function
}
\description{
Compute the approximation to the sin() function.
}
\details{
This function computes the approximation to the Sine function.

If the input data contains NaN value, the function will print warning to remind.

If the input data contains Non-numeric value, the function will print warning to remind.
}
\examples{
fn_sin(c(1:10), 5)

}
